+++ demo/backend/app.tsx
@@ -86,7 +86,6 @@ export function startDemoServer(options?: { port?: number }) {
 
       const handledHeadTagRequest = await handleHeadTagRequest(request, {
         headTags,
-        headBasePath: '/head-api',
       });
 
       if (handledHeadTagRequest.matched) {
@@ -101,7 +100,6 @@ export function startDemoServer(options?: { port?: number }) {
       const handled = await handleRequest(request, {
         routeManifest,
         headTags,
-        headBasePath: '/head-api',
         html: {
           template: htmlTemplate,
         },