+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @richie-router/react
 
+## 0.1.6
+
+### Patch Changes
+
+- add passthrough routes (that go directly to the API) and move headBasePath to the RouterSchema
+- Updated dependencies
+  - @richie-router/core@0.1.5
+
 ## 0.1.5
 
 ### Patch Changes